Change Map on Set Lead to the same type id and idGenerator

+++ src/main/java/school/faang/sprint1/task_BJS2_56911/Product.java
@@ -4,14 +4,11 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
-import java.util.HashMap;
-import java.util.Map;
-
 @EqualsAndHashCode
 @Getter
 @ToString
 public class Product {
-    private static int idGenerator;
+    private static long idGenerator;
     private long id;
     private String name;
     private Category category;
